Remove redundant code.
