+++ js/src/meta-boxes/channel-visibility/google-ads-promo-cta.js
@@ -8,24 +8,16 @@ import { Flex, FlexBlock } from '@wordpress/components';
  * Internal dependencies
  */
 import AppButton from '~/components/app-button';
-import { getGetStartedUrl } from '~/utils/urls';
+import GetStartedCTA from './get-started-cta';
 import { addBaseEventProperties } from '~/utils/tracks';
 
-const GET_STARTED_URL = getGetStartedUrl();
 const BASE_EVENT_PROPS = addBaseEventProperties( {} );
 
-const GoogleAdsPromoCTA = () => {
+const GoogleAdsPromoCTA = ( { onDismiss } ) => {
 	return (
 		<Flex className="gla-channel-visibility-google-ads-promo-cta" gap={ 4 }>
 			<FlexBlock>
-				<AppButton
-					href={ GET_STARTED_URL }
-					eventName="gla_google_ads_promo_get_started_click"
-					eventProps={ BASE_EVENT_PROPS }
-					isSecondary
-				>
-					{ __( 'Get started', 'google-listings-and-ads' ) }
-				</AppButton>
+				<GetStartedCTA />
 			</FlexBlock>
 
 			<FlexBlock>
@@ -33,6 +25,7 @@ const GoogleAdsPromoCTA = () => {
 					eventName="gla_google_ads_promo_dismiss_click"
 					eventProps={ BASE_EVENT_PROPS }
 					isTertiary
+					onClick={ onDismiss }
 				>
 					{ __( 'Dismiss', 'google-listings-and-ads' ) }
 				</AppButton>
