+++ js/src/meta-boxes/channel-visibility/google-ads-promo-setup-completed.js
@@ -7,13 +7,7 @@ import { Flex, FlexItem, SelectControl } from '@wordpress/components';
 /**
  * Internal dependencies
  */
-import AppButton from '~/components/app-button';
-import { getGetStartedUrl } from '~/utils/urls';
 import googleLogoURL from '~/images/logo/gogole-g-logo.svg';
-import { addBaseEventProperties } from '~/utils/tracks';
-
-const GET_STARTED_URL = getGetStartedUrl();
-const BASE_EVENT_PROPS = addBaseEventProperties( {} );
 
 const GoogleAdsPromoSetupCompleted = () => {
 	return (
@@ -57,14 +51,6 @@ const GoogleAdsPromoSetupCompleted = () => {
 					] }
 					__nextHasNoMarginBottom
 				/>
-				<AppButton
-					href={ GET_STARTED_URL }
-					eventName="gla_google_ads_promo_get_started_click"
-					eventProps={ BASE_EVENT_PROPS }
-					isSecondary
-				>
-					{ __( 'Get started', 'google-listings-and-ads' ) }
-				</AppButton>
 			</FlexItem>
 		</Flex>
 	);
