+++ js/src/meta-boxes/channel-visibility/google-ads-promo.scss
@@ -12,4 +12,8 @@
 		padding: $grid-unit-20;
 		align-items: flex-start;
 	}
+
+	.gla-channel-visibility-get-started-is-dismissed {
+		margin-left: auto;
+	}
 }
